fix: integrated ciphers into codecs
