+++ core/src/main/java/de/pxav/kelp/core/connect/KelpConnectVersionTemplate.java
@@ -1,8 +1,6 @@
 package de.pxav.kelp.core.connect;
 
 import de.pxav.kelp.core.application.KelpVersionTemplate;
-import de.pxav.kelp.core.connect.connection.IConnectionDecrypter;
-import de.pxav.kelp.core.connect.connection.IConnectionEncrypter;
 import de.pxav.kelp.core.connect.packet.IPacketDecoder;
 import de.pxav.kelp.core.connect.packet.IPacketEncoder;
 import de.pxav.kelp.core.connect.packet.PacketRegistry;
@@ -15,11 +13,7 @@ import javax.crypto.Cipher;
 @KelpVersionTemplate
 public abstract class KelpConnectVersionTemplate {
 
-  public abstract IConnectionEncrypter newEncrypter(Cipher encrypter);
+  public abstract IPacketEncoder newPacketEncoder(PacketRegistry registry, Cipher encrypter);
 
-  public abstract IConnectionDecrypter newDecrypter(Cipher decrypter);
-
-  public abstract IPacketEncoder newPacketEncoder(PacketRegistry registry);
-
-  public abstract IPacketDecoder newPacketDecoder(PacketRegistry registry);
+  public abstract IPacketDecoder newPacketDecoder(PacketRegistry registry, Cipher decrypter);
 }
