--- core/src/main/java/de/pxav/kelp/core/connect/connection/IConnectionEncrypter.java
@@ -1,7 +1,0 @@
-package de.pxav.kelp.core.connect.connection;
-
-/**
- * @author Etrayed
- */
-public interface IConnectionEncrypter {
-}
