+++ v1_8_implementation/src/main/java/de/pxav/kelp/implementation1_8/connect/VersionedKelpConnect.java
@@ -2,14 +2,10 @@ package de.pxav.kelp.implementation1_8.connect;
 
 import com.google.inject.Singleton;
 import de.pxav.kelp.core.connect.KelpConnectVersionTemplate;
-import de.pxav.kelp.core.connect.connection.IConnectionDecrypter;
-import de.pxav.kelp.core.connect.connection.IConnectionEncrypter;
 import de.pxav.kelp.core.connect.packet.IPacketDecoder;
 import de.pxav.kelp.core.connect.packet.IPacketEncoder;
 import de.pxav.kelp.core.connect.packet.PacketRegistry;
 import de.pxav.kelp.core.version.Versioned;
-import de.pxav.kelp.implementation1_8.connect.connection.ConnectionDecrypter;
-import de.pxav.kelp.implementation1_8.connect.connection.ConnectionEncrypter;
 import de.pxav.kelp.implementation1_8.connect.packet.PacketDecoder;
 import de.pxav.kelp.implementation1_8.connect.packet.PacketEncoder;
 
@@ -23,22 +19,12 @@ import javax.crypto.Cipher;
 public class VersionedKelpConnect extends KelpConnectVersionTemplate {
 
   @Override
-  public IConnectionEncrypter newEncrypter(Cipher encrypter) {
-    return new ConnectionEncrypter(encrypter);
+  public IPacketEncoder newPacketEncoder(PacketRegistry registry, Cipher encrypter) {
+    return new PacketEncoder(registry, encrypter);
   }
 
   @Override
-  public IConnectionDecrypter newDecrypter(Cipher decrypter) {
-    return new ConnectionDecrypter(decrypter);
-  }
-
-  @Override
-  public IPacketEncoder newPacketEncoder(PacketRegistry registry) {
-    return new PacketEncoder(registry);
-  }
-
-  @Override
-  public IPacketDecoder newPacketDecoder(PacketRegistry registry) {
-    return new PacketDecoder(registry);
+  public IPacketDecoder newPacketDecoder(PacketRegistry registry, Cipher decrypter) {
+    return new PacketDecoder(registry, decrypter);
   }
 }
